Refactor CND backend modules into municipal package
