+++ backend/cnds/municipal/__init__.py
@@ -1,0 +1,4 @@
+"""Municipal CND automation package."""
+from .routes import router
+
+__all__ = ["router"]
